V1

+++ app/surveymodule/[accessKey]/page.tsx
@@ -22,6 +22,7 @@ import AutoAwesomeIcon from '@mui/icons-material/AutoAwesome';
 import Tooltip from '@mui/material/Tooltip';
 import {CopyToClipboard} from 'react-copy-to-clipboard';
 import ContentCopyIcon from '@mui/icons-material/ContentCopy';
+import HelpOutlineIcon from '@mui/icons-material/HelpOutline';
 
 interface SurveyPageProps {
   params: {
@@ -142,7 +143,9 @@ export default function QuestionsPage({ params }: SurveyPageProps) {
   }, []);
 
   // Deletion
-  const handleDeleteSurvey = async (SurveyID: string) => {
+  const handleDeleteSurvey = async (SurveyID: string, SurveyTitle: string) => {
+    const userConfirmed = confirm(`Are you sure you want to delete the survey "${SurveyTitle}"?`);
+    if (!userConfirmed) return;
     try {
       setIsLoading(true);
       await deleteSurvey(params.accessKey, SurveyID);
@@ -215,11 +218,27 @@ export default function QuestionsPage({ params }: SurveyPageProps) {
                 </div>
               </div>
               <div className={styles.sidebarFormBit}>
-                <label className={styles.sidebarLabel}>Required No. of Sessions</label>
+                <div style={{display: "flex", flexDirection: "row"}}>
+                  <label className={styles.sidebarLabel}>Required No. of Sessions</label>
+                  <Tooltip title="Number of sessions required to complete the survey"
+                    placement="top" arrow
+                    slotProps={{ tooltip: { sx: { fontSize: '0.8em' } } }}
+                  >
+                    <HelpOutlineIcon style={{color: "#E07954", fontSize: "1.3rem", marginLeft: '2%', marginTop: '0.5%'}} />
+                  </Tooltip>
+                </div>
                 <input type="number" required min="1" name="Sessions" className={styles.sidebarTextField} />
               </div>
               <div className={styles.sidebarFormBit}>
-                <label className={styles.sidebarLabel}>Minimum Interval (in hours)</label>
+                <div style={{display: "flex", flexDirection: "row"}}>
+                  <label className={styles.sidebarLabel}>Minimum Interval (in hours)</label>
+                  <Tooltip title="Minimum time gap to answer sessions"
+                    placement="top" arrow
+                    slotProps={{ tooltip: { sx: { fontSize: '0.8em' } } }}
+                  >
+                    <HelpOutlineIcon style={{color: "#E07954", fontSize: "1.3rem", marginLeft: '2%', marginTop: '0.5%'}} />
+                  </Tooltip>
+                </div>
                 <input type="number" required min="0" step="0.5" name="Interval" className={styles.sidebarTextField} />
               </div>
               {error && <p className={styles.errorMessage}>{error}</p>}
@@ -280,7 +299,7 @@ export default function QuestionsPage({ params }: SurveyPageProps) {
                 </Tooltip>
 
                 <Tooltip title="Delete" arrow placement="top">
-                <button onClick={() => handleDeleteSurvey(survey.id)}>
+                <button onClick={() => handleDeleteSurvey(survey.id, survey.data.Title)}>
                   <DeleteOutlineIcon sx={{ fontSize: 30, color: '#E07961' }}/>
                 </button>
                 </Tooltip>
